@@ -1,9 +1,9 @@
 {
-  "name": "tech-test",
-  "version": "0.0.0",
+  "name": "angular-todo",
+  "version": "0.1.0",
   "scripts": {
     "ng": "ng",
-    "start": "set NODE_OPTIONS=--openssl-legacy-provider && ng serve --open",
+    "start": "ng serve --open",
     "build": "ng build",
     "test": "set NODE_OPTIONS=--openssl-legacy-provider && ng test",
     "lint": "ng lint",
@@ -12,16 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.10",
-    "@angular/common": "~12.2.10",
-    "@angular/compiler": "~12.2.10",
-    "@angular/core": "~12.2.10",
-    "@angular/forms": "~12.2.10",
-    "@angular/localize": "~12.2.10",
-    "@angular/platform-browser": "~12.2.10",
-    "@angular/platform-browser-dynamic": "~12.2.10",
-    "@angular/router": "~12.2.10",
-    "@ng-bootstrap/ng-bootstrap": "^10.0.0",
+    "@angular/animations": "^15.1.1",
+    "@angular/common": "^15.1.1",
+    "@angular/compiler": "^15.1.1",
+    "@angular/core": "^15.1.1",
+    "@angular/forms": "^15.1.1",
+    "@angular/localize": "^15.1.1",
+    "@angular/platform-browser": "^15.1.1",
+    "@angular/platform-browser-dynamic": "^15.1.1",
+    "@angular/router": "^15.1.1",
+    "@ng-bootstrap/ng-bootstrap": "^14.0.1",
+    "@popperjs/core": "^2.11.6",
     "bootstrap": "^5.2.3",
     "bootstrap-icons": "^1.10.3",
     "json-server": "^0.17.0",
@@ -29,14 +30,14 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.10",
-    "@angular/cli": "~12.2.10",
-    "@angular/compiler-cli": "~12.2.10",
-    "@angular/language-service": "~12.2.10",
+    "@angular-devkit/build-angular": "^15.1.2",
+    "@angular/cli": "^15.1.2",
+    "@angular/compiler-cli": "^15.1.1",
+    "@angular/language-service": "^15.1.1",
     "@types/jasmine": "~3.10.0",
     "@types/jasminewd2": "~2.0.10",
     "@types/node": "~16.11.1",
-    "codelyzer": "^6.0.2",
+    "codelyzer": "^0.0.28",
     "jasmine-core": "~3.10.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.3.4",
@@ -48,6 +49,6 @@
     "ts-node": "^10.2.1",
     "tslib": "^2.4.1",
     "tslint": "~5.20.1",
-    "typescript": "<4.4"
+    "typescript": "4.9.4"
   }
 }
